indexedDB implementation _almost_ works
logic.js is a complete mess though.  Currently, nothing shows because there is no switch to load the boilerplate checklist the first time through, and it doesn't makes sense to reload it every time, because that defeats the purpose of persistence anyway.

The draggable code is back, but it doesn't work either. That fix is still coming.

Use this code only for experimenting